op refactory and testing finished

+++ auth/models/operations/delete-session.js
@@ -1,23 +1,21 @@
 import Session from 'auth/models/schemas/session.js'
 
 export default async(info) => {
+  if(typeof info !== 'object' || info === null){
+    let err = new Error('invalid argument')
+    err.status = 400
+    throw err
+  }
+
   try{
-    try{
-      await Session.destroy({
-        where: {
-          sessionId: info.sessionId,
-        },
-      })
-    }catch(err){
-      err = new Error('session deletion failed')
-      err.status = 500
-      throw err
-    }
+    await Session.destroy({
+      where: {
+        sessionId: info.sessionId,
+      },
+    })
   }catch(err){
-    if(typeof err.status !== 'number'){
-      err = new Error('delete-session failed')
-      err.status = 500
-    }
+    err = new Error('deleting session failed')
+    err.status = 500
     throw err
   }
 }
